+++ src/main/java/ru/sportlive/mvp/controller/UserController.java
@@ -3,6 +3,7 @@ package ru.sportlive.mvp.controller;
 
 
 import io.swagger.v3.oas.annotations.Operation;
+import jakarta.servlet.http.HttpSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -14,7 +15,7 @@ import ru.sportlive.mvp.services.UserService;
 import java.util.List;
 
 @RestController
-@RequestMapping("/user")
+@RequestMapping("/api/user")
 public class UserController {
 
     @Autowired
@@ -22,21 +23,27 @@ public class UserController {
 
     @PostMapping("/")
     @Operation(summary = "Добавляет нового пользователя")
-    public ResponseEntity<User>addNewUser (@RequestBody UsersDTO usersDTO){
+    public ResponseEntity<User>addNewUser (@RequestBody UsersDTO usersDTO, HttpSession httpSession){
         try {
             User user = userService.addUser(usersDTO.getName(), usersDTO.getSurname(), usersDTO.getHeight(),usersDTO.getWeight());
+            httpSession.setAttribute("userId",user.getId());
             return new ResponseEntity<>(user, HttpStatus.OK);
         } catch (Exception e) {
             throw new RuntimeException(e.getMessage());
         }
     }
-    @DeleteMapping("/{id}")
+    @DeleteMapping("/")
     @Operation(summary = "Удаляет пользователя по id")
-    public ResponseEntity<User>deleteUser(@PathVariable Integer id) throws Exception {
+    public ResponseEntity<User>deleteUser(HttpSession httpSession ) throws Exception {
+        Integer id = (Integer) httpSession.getAttribute("userId");
+        if (id == null){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
         User user = userService.deleteUser(id);
+
         return new ResponseEntity<>(user,HttpStatus.OK);
     }
-    @GetMapping("/")
+    @GetMapping("/all")
     @Operation(summary = "Выводит всех пользователей")
     public ResponseEntity<List<User>>getAllUser(){
         List<User>users = userService.getAllUsers();
@@ -50,10 +57,25 @@ public class UserController {
         return new ResponseEntity<>(user,HttpStatus.OK);
     }
 
+    @Operation(summary = "Выводит авторизированного пользователя")
+    @GetMapping("/")
+    public ResponseEntity<User>getUserAuth(HttpSession httpSession){
+        Integer id = (Integer) httpSession.getAttribute("userId");
+        if (id == null){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+        User user =userService.getUser(id);
+        return new ResponseEntity<>(user,HttpStatus.OK);
+    }
+
     @Operation(summary = "Обновления данных у пользователя")
-    @PutMapping("/{user_id}")
-    public ResponseEntity<User>updateUser(@PathVariable Integer user_id,@RequestBody UsersDTO usersDTO){
-        User user = userService.getUser(user_id);
+    @PutMapping("/")
+    public ResponseEntity<User>updateUser(@RequestBody UsersDTO usersDTO,HttpSession httpSession){
+        Integer id = (Integer) httpSession.getAttribute("userId");
+        if (id == null){
+            return new ResponseEntity<>(HttpStatus.UNAUTHORIZED);
+        }
+        User user = userService.getUser(id);
         user = userService.updateToUser(user, usersDTO);
         return new ResponseEntity<>(user,HttpStatus.OK);
     }
